@@ -17,45 +17,45 @@
     "preinstall": "node scripts/initialSetting.js"
   },
   "dependencies": {
-    "@astrojs/mdx": "^3.1.7",
-    "@astrojs/svelte": "^5.7.1",
-    "@inquirer/prompts": "^6.0.1",
-    "@mkkellogg/gaussian-splats-3d": "^0.4.5",
-    "@threlte/core": "^7.3.1",
-    "@threlte/extras": "^8.12.0",
+    "@astrojs/mdx": "^4.0.7",
+    "@astrojs/svelte": "^7.0.4",
+    "@inquirer/prompts": "^7.2.3",
+    "@mkkellogg/gaussian-splats-3d": "^0.4.6",
+    "@threlte/core": "^8.0.1",
+    "@threlte/extras": "^9.0.1",
     "@types/fs-extra": "^11.0.4",
-    "astro": "^4.15.9",
+    "astro": "^5.1.9",
     "camera-controls": "^2.9.0",
-    "fs-extra": "^11.2.0",
+    "fs-extra": "^11.3.0",
+    "lenis": "^1.1.20",
     "sharp": "^0.33.5",
-    "svelte": "^4.2.19",
-    "svelte-tweakpane-ui": "^1.5.0",
-    "three": "^0.169.0",
+    "svelte": "^5.1.16",
+    "svelte-tweakpane-ui": "^1.5.1",
+    "three": "^0.172.0",
     "tree-node-cli": "^1.6.0",
-    "typescript": "^5.6.2"
+    "typescript": "^5.7.3"
   },
   "devDependencies": {
-    "@markuplint/astro-parser": "^4.6.7",
-    "@types/three": "^0.169.0",
-    "@typescript-eslint/eslint-plugin": "^8.8.0",
-    "@typescript-eslint/parser": "^8.8.0",
-    "eslint": "^9.11.1",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-astro": "^1.2.4",
-    "eslint-plugin-jsx-a11y": "^6.10.0",
-    "glob": "^11.0.0",
-    "markuplint": "^4.9.4",
-    "node-html-parser": "^6.1.13",
-    "postcss-html": "^1.7.0",
-    "prettier": "^3.3.3",
+    "@markuplint/astro-parser": "^4.6.14",
+    "@types/three": "^0.172.0",
+    "@typescript-eslint/eslint-plugin": "^8.21.0",
+    "@typescript-eslint/parser": "^8.21.0",
+    "eslint": "^9.18.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-astro": "^1.3.1",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "glob": "^11.0.1",
+    "markuplint": "^4.11.4",
+    "node-html-parser": "^7.0.1",
+    "postcss-html": "^1.8.0",
+    "prettier": "^3.4.2",
     "prettier-plugin-astro": "^0.14.1",
-    "prettier-plugin-svelte": "^3.2.7",
-    "sass": "^1.79.4",
-    "stylelint": "^16.9.0",
+    "prettier-plugin-svelte": "^3.3.3",
+    "sass": "^1.83.4",
+    "stylelint": "^16.13.2",
     "stylelint-config-html": "^1.1.0",
-    "stylelint-config-recess-order": "^5.1.1",
-    "stylelint-config-standard-scss": "^13.1.0",
-    "svg-sprite": "^2.0.4",
-    "vituum": "^1.1.1"
+    "stylelint-config-recess-order": "^6.0.0",
+    "stylelint-config-standard-scss": "^14.0.0",
+    "svg-sprite": "^2.0.4"
   }
 }
